rpc: Combine BlockMetaService and TransactionStatusService
