+++ ledger-tool/src/ledger_utils.rs
@@ -27,9 +27,7 @@ use {
     },
     solana_measure::measure_time,
     solana_pubkey::Pubkey,
-    solana_rpc::{
-        block_meta_service::BlockMetaService, transaction_status_service::TransactionStatusService,
-    },
+    solana_rpc::transaction_status_service::TransactionStatusService,
     solana_runtime::{
         accounts_background_service::{
             AbsRequestHandlers, AccountsBackgroundService, PrunedBanksRequestHandler,
@@ -294,57 +292,41 @@ pub fn load_and_process_ledger(
 
     let enable_rpc_transaction_history = arg_matches.is_present("enable_rpc_transaction_history");
 
-    let (
-        transaction_status_sender,
-        transaction_status_service,
-        block_meta_sender,
-        block_meta_service,
-    ) = if geyser_plugin_active || enable_rpc_transaction_history {
-        // Need Primary (R/W) access to insert transaction and rewards data;
-        // obtain Primary access if we do not already have it
-        let write_blockstore = if enable_rpc_transaction_history && !blockstore.is_primary_access()
-        {
-            Arc::new(open_blockstore(
-                blockstore.ledger_path(),
-                arg_matches,
-                AccessType::PrimaryForMaintenance,
-            ))
+    let (transaction_status_sender, transaction_status_service) =
+        if geyser_plugin_active || enable_rpc_transaction_history {
+            // Need Primary (R/W) access to insert transaction and rewards data;
+            // obtain Primary access if we do not already have it
+            let write_blockstore =
+                if enable_rpc_transaction_history && !blockstore.is_primary_access() {
+                    Arc::new(open_blockstore(
+                        blockstore.ledger_path(),
+                        arg_matches,
+                        AccessType::PrimaryForMaintenance,
+                    ))
+                } else {
+                    blockstore.clone()
+                };
+
+            let (transaction_status_sender, transaction_status_receiver) = unbounded();
+            let transaction_status_service = TransactionStatusService::new(
+                transaction_status_receiver,
+                Arc::default(),
+                enable_rpc_transaction_history,
+                transaction_notifier,
+                write_blockstore.clone(),
+                arg_matches.is_present("enable_extended_tx_metadata_storage"),
+                tss_exit.clone(),
+            );
+
+            (
+                Some(TransactionStatusSender {
+                    sender: transaction_status_sender,
+                }),
+                Some(transaction_status_service),
+            )
         } else {
-            blockstore.clone()
+            (transaction_status_sender, None)
         };
-
-        let (transaction_status_sender, transaction_status_receiver) = unbounded();
-        let transaction_status_service = TransactionStatusService::new(
-            transaction_status_receiver,
-            Arc::default(),
-            enable_rpc_transaction_history,
-            transaction_notifier,
-            write_blockstore.clone(),
-            arg_matches.is_present("enable_extended_tx_metadata_storage"),
-            tss_exit.clone(),
-        );
-
-        let (block_meta_sender, block_meta_receiver) = unbounded();
-        // Nothing else will be interacting with max_complete_rewards_slot
-        let max_complete_rewards_slot = Arc::default();
-        let block_meta_service = BlockMetaService::new(
-            block_meta_receiver,
-            write_blockstore,
-            max_complete_rewards_slot,
-            exit.clone(),
-        );
-
-        (
-            Some(TransactionStatusSender {
-                sender: transaction_status_sender,
-            }),
-            Some(transaction_status_service),
-            Some(block_meta_sender),
-            Some(block_meta_service),
-        )
-    } else {
-        (transaction_status_sender, None, None, None)
-    };
 
     let (bank_forks, leader_schedule_cache, starting_snapshot_hashes, ..) =
         bank_forks_utils::load_bank_forks(
@@ -353,7 +335,7 @@ pub fn load_and_process_ledger(
             account_paths,
             &snapshot_config,
             &process_options,
-            block_meta_sender.as_ref(),
+            None, // block_meta_sender, delete in next commit
             None, // Maybe support this later, though
             accounts_update_notifier,
             exit.clone(),
@@ -438,7 +420,7 @@ pub fn load_and_process_ledger(
         &leader_schedule_cache,
         &process_options,
         transaction_status_sender.as_ref(),
-        block_meta_sender.as_ref(),
+        None, // block_meta_sender, delete in next commit
         None, // entry_notification_sender
         Some(&snapshot_controller),
     )
@@ -453,9 +435,6 @@ pub fn load_and_process_ledger(
     accounts_hash_verifier.join().unwrap();
     if let Some(service) = transaction_status_service {
         service.quiesce_and_join_for_tests(tss_exit);
-    }
-    if let Some(service) = block_meta_service {
-        service.join().unwrap();
     }
 
     result
