+++ rpc/src/lib.rs
@@ -1,5 +1,4 @@
 #![allow(clippy::arithmetic_side_effects)]
-pub mod block_meta_service;
 mod cluster_tpu_info;
 pub mod filter;
 pub mod max_slots;
