+++ rpc/src/transaction_status_service.rs
@@ -6,9 +6,11 @@ use {
         blockstore::{Blockstore, BlockstoreError},
         blockstore_processor::{TransactionStatusBatch, TransactionStatusMessage},
     },
+    solana_runtime::bank::{Bank, KeyedRewardsAndNumPartitions},
     solana_svm::transaction_commit_result::CommittedTransaction,
     solana_transaction_status::{
-        extract_and_fmt_memos, map_inner_instructions, Reward, TransactionStatusMeta,
+        extract_and_fmt_memos, map_inner_instructions, Reward, RewardsAndNumPartitions,
+        TransactionStatusMeta,
     },
     std::{
         sync::{
@@ -230,10 +232,44 @@ impl TransactionStatusService {
                     blockstore.write_batch(status_and_memos_batch)?;
                 }
             }
-            TransactionStatusMessage::Freeze(slot) => {
-                max_complete_transaction_status_slot.fetch_max(slot, Ordering::SeqCst);
+            TransactionStatusMessage::Freeze(bank) => {
+                Self::write_block_meta(&bank, blockstore)?;
+                max_complete_transaction_status_slot.fetch_max(bank.slot(), Ordering::SeqCst);
             }
         }
+        Ok(())
+    }
+
+    fn write_block_meta(bank: &Bank, blockstore: &Blockstore) -> Result<(), BlockstoreError> {
+        let slot = bank.slot();
+
+        blockstore.set_block_time(slot, bank.clock().unix_timestamp)?;
+        blockstore.set_block_height(slot, bank.block_height())?;
+
+        let rewards = bank.get_rewards_and_num_partitions();
+        if rewards.should_record() {
+            let KeyedRewardsAndNumPartitions {
+                keyed_rewards,
+                num_partitions,
+            } = rewards;
+            let rewards = keyed_rewards
+                .into_iter()
+                .map(|(pubkey, reward_info)| Reward {
+                    pubkey: pubkey.to_string(),
+                    lamports: reward_info.lamports,
+                    post_balance: reward_info.post_balance,
+                    reward_type: Some(reward_info.reward_type),
+                    commission: reward_info.commission,
+                })
+                .collect();
+            let blockstore_rewards = RewardsAndNumPartitions {
+                rewards,
+                num_partitions,
+            };
+
+            blockstore.write_rewards(slot, blockstore_rewards)?;
+        }
+
         Ok(())
     }
